chore: fix tests and write interactions
